Removed Lists containing other domain objects.
- The reason for this is that the lists are always empty when we retrieve the data from Mongo.
- In an object-oriented way of thinking this makes sense. (A customer owns a list of accounts and accounts own a list of transactions).
- However this means that if we wanted to print a single customer we would have to make recursive calls to every list it has.

In a REST API it makes no sense to get every single piece of data from just calling /customer/{customer_id} but if we keep it like this and populate it you would get all the customer details, plus a list of all accounts, and then within that you would also have a list of all transactions. That would make just getting the customer details a massive request so it is better to break this relationship apart.

What we do is break this down into separate requests, using the previous 'owners' ID as a reference back and keeping it linked without having them nested.